Modularized all codes
