+++ dirtyreact-app/src/hiring.js
@@ -1,0 +1,5 @@
+import React from 'react'
+export const Hiring = () => 
+<div>
+  <p>The library is hiring. Go to www.library.com/jobs for more</p>
+</div>
