+++ dirtyreact-app/src/nothiring.js
@@ -1,0 +1,5 @@
+import React from 'react'
+export const NotHiring = () =>
+    <div>
+      <p>The Library is not hiring. Check back later for more.</p>
+    </div>
